Add warnings and options to unit_testing.sh
Adds a warning to unit_testing.sh informing the user that
the contents of their /opt folder will be deleted by the test.
Also adds option parsing, -y for skipping the prompt and -d
for adding xtrace debugging, as well as -h for displaying help.

+++ bootstrap/vagrant-onap/unit_testing.sh
@@ -1,6 +1,40 @@
 #!/bin/bash
 
-set -o xtrace
+run=false
+
+while getopts "ydh" OPTION; do
+    case $OPTION in
+        y)
+            run=true
+            ;;
+        d)
+            set -o xtrace
+            ;;
+        h)
+            echo "Usage: unit_testing.sh [-y] [-d] [-h]"
+            echo "Optional arguments:"
+            echo "    -y    Skips warning prompt."
+            echo "    -d    Shows debugging output."
+            echo "    -h    Shows help about this program."
+            exit 0
+            ;;
+    esac
+done
+
+if [ "$run" = false ] ; then
+    while true; do
+        echo "Warning: This test script will delete the contents of your /opt folder."
+        read -p "Would you like to continue? [y]es/[n]o" yn
+        case $yn in
+            [Yy]*)
+                break
+                ;;
+            [Nn]*)
+                exit
+                ;;
+        esac
+    done
+fi
 
 rm -rf /opt/
 rm -rf /root/.m2/
